refactor: update Document import paths to ile subdirectory
